Improved security, decent_exposure, revisions work

+++ app/controllers/phrases_controller.rb
@@ -2,11 +2,13 @@
 class PhrasesController < ApplicationController
   expose(:phrase)
   expose(:phrases) {
-  	p = Phrase.search {fulltext params[:search] do
+  	p = Phrase.search {
+      fulltext params[:search] do
 	  	phrase_fields :title => 5.0
 	  	phrase_fields :description => 3.0
 	  	phrase_slop 1
 	  	end
+      with :published, true
 	  }
 	p.results
   }
@@ -14,31 +16,43 @@ class PhrasesController < ApplicationController
   respond_to :html
 
   def create
-  	if phrase.save
-      revision = Revision.from_phrase(phrase)
-      revision.author_id = current_author.id
-      revision.save
-      flash[:notice] = "Dodano frazę."
+    if !author_signed_in?
+      redirect_to new_author_session_path
+    else
+    	if phrase.save
+        revision = Revision.from_phrase(phrase)
+        revision.author_id = current_author.id
+        revision.save
+        flash[:notice] = "Dodano frazę."
+      end
+    	respond_with phrase
     end
-  	respond_with phrase
   end
 
   def update
-    if phrase.save
-      revision = Revision.from_phrase(phrase)
-      revision.author_id = current_author.id
-      revision.save
-      flash[:notice] = "Zaktualizowano frazę."
+    if !author_signed_in?
+      redirect_to new_author_session_path
+    else
+      if phrase.save
+        revision = Revision.from_phrase(phrase)
+        revision.author_id = current_author.id
+        revision.save
+        flash[:notice] = "Zaktualizowano frazę."
+      end
+      respond_with phrase
     end
-    respond_with phrase
   end
 
   def destroy
-    revision = Revision.from_phrase(phrase)
-    revision.author_id = current_author.id
-    revision.save
-    phrase.destroy
-    flash[:notice] = "Usunięto frazę."
-    redirect_to homepage_path
+    if !author_signed_in?
+      redirect_to new_author_session_path
+    else
+      revision = Revision.from_phrase(phrase)
+      revision.author_id = current_author.id
+      revision.save
+      phrase.destroy
+      flash[:notice] = "Usunięto frazę."
+      redirect_to homepage_path
+    end
   end
 end
